buttonList - baloon
δημιούργησα ένα νέο obj για το buttonList στο epiloges.
πρόσθεσα και τις αλλαγές σου για το baloon στο epiloges
